+++ wemake_python_styleguide/visitors/ast/complexity/counts.py
@@ -6,10 +6,13 @@ from typing import ClassVar, DefaultDict, List, Union
 
 from typing_extensions import final
 
+from wemake_python_styleguide.constants import MAX_LEN_YIELD_TUPLE
+
 from wemake_python_styleguide.logics.functions import is_method
 from wemake_python_styleguide.logics.nodes import get_parent
 from wemake_python_styleguide.types import AnyFunctionDef, AnyImport
 from wemake_python_styleguide.violations.complexity import (
+    TooLongYieldTupleViolation,
     TooManyConditionsViolation,
     TooManyDecoratorsViolation,
     TooManyElifsViolation,
@@ -252,4 +255,34 @@ class TryExceptVisitor(BaseNodeVisitor):
 
         """
         self._check_except_count(node)
+        self.generic_visit(node)
+
+
+@final
+@alias('visit_any_function', (
+    'visit_FunctionDef',
+    'visit_AsyncFunctionDef',
+))
+class YieldTupleVisitor(BaseNodeVisitor):
+    """Finds too long ``tuples`` in ``yield`` expressions."""
+
+    def _check_yield_values(self, node: AnyFunctionDef) -> None:
+        for sub_node in ast.walk(node):
+            if isinstance(sub_node, ast.Yield):
+                if sub_node.value and isinstance(sub_node.value, ast.Tuple):
+                    yield_list = [x for x in sub_node.value.elts]
+                    if len(yield_list) > MAX_LEN_YIELD_TUPLE:
+                        self.add_violation(
+                            TooLongYieldTupleViolation(sub_node, text=len(yield_list))
+                        )
+
+    def visit_any_function(self, node: AnyFunctionDef) -> None:
+        """
+        Helper to get all ``yield`` nodes in a function at once.
+
+        Raises:
+            TooLongYieldTupleViolation
+
+        """
+        self._check_yield_values(node)
         self.generic_visit(node)