+++ CHANGELOG.md
@@ -13,6 +13,7 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to over-use strings
 - Forbids to use `breakpoint` function
 - Adds `flake8-rst-docstrings` as a dependency
+- Limits yield tuple lengths
 
 
 ## 0.9.0